+++ src/modules/book/BookData.h
@@ -5,11 +5,10 @@
 
 namespace faker
 {
-extern const std::array<std::string_view, 100> authors;
-extern const std::array<std::string_view, 3> bookFormats;
+extern const std::array<std::string_view, 125> authors;
+extern const std::array<std::string_view, 4> bookFormats;
 extern const std::array<std::string_view, 25> bookGenres;
 extern const std::array<std::string_view, 263> publishers;
-extern const std::array<std::string_view, 19> bookSeries;
-extern const std::array<std::string_view, 100> titles;
-extern const std::array<std::string_view, 15> translators;
+extern const std::array<std::string_view, 34> bookSeries;
+extern const std::array<std::string_view, 102> titles;
 }